BugFix - Only owner of the image to edit/delete the image

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,18 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
+        User loggedInUser = (User) session.getAttribute("loggeduser");
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if (isOwnerOfImage(image.getUser().getId(), loggedInUser.getId())) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        String error = "Only the owner of the image can edit the image";
+        model.addAttribute("editError", error);
+        return "images/image";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +145,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+        if (isOwnerOfImage(image.getUser().getId(), loggedInUser.getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        model.addAttribute("image", image);
+        String error = "Only the owner of the image can delete the image";
+        model.addAttribute("deleteError", error);
+        return "images/image";
+    }
+
+    private boolean isOwnerOfImage(Integer loggedOutUserId, Integer loggedInUserId) {
+        return Objects.equals(loggedOutUserId, loggedInUserId);
     }
